+++ lib/toy/extensions/uuid.rb
@@ -1,16 +1,14 @@
 module Toy
   module Extensions
     module UUID
-      def store_default
-        new
-      end
-
       def to_store(value, *)
+        return nil if value.nil?
         return value if value.is_a?(self)
         new(value)
       end
 
       def from_store(value, *)
+        return nil if value.nil?
         return value if value.is_a?(self)
         new(value)
       end